Adjust letter positions, and serve only under /v2
A couple of letters ('D' and 'P') didn't look quite right previously, so
I've switched to individual, per-letter position adjustments.  Code is a bit
more complicated, but it means that I don't have to compromise positioning
for *anyone*.

Also, it was an oversight to change existing avatar rendering, so I've now
changed things so that you get the exact same avatars you always have on the
existing URLs, and you need to prepend `/v2` to the path to get the new
Roboto ones (so `/v2/letters/Q/<etc>`).