Dockerfile: change base image to debian bookworm
This drops Alpine as the base O/S and replaces the equivalent
dependencies. It is easier for us to monitor security notices when
images are built from a common base ie. in line with the Discourse
application image built out from github.com/discourse/discourse_docker.

Additionally, we don't need to build the runtime dependencies from
source now that newer stable version are available from upstream Deb
repositories and it greatly simplifies the image build.