fix COPY of lib directory
We must specify a seperate COPY command for the application files to end
up in the /var/www/letter-avatars/lib directory. Otherwise, these files
end up in the root of the directory and config.ru cannot require them.

Also made usage of COPY consistent rather than a mix of ADD and COPY.

Follow-up to 6e3cb40.